[CHORE] Rename lsp config file to avoid name conflicts
